user_service/requirements.txt
asgiref==3.8.1
Django==4.2.16
sqlparse==0.5.3
typing_extensions==4.12.2
psycopg2==2.9.10
djangorestframework==3.15.2
pytest
pytest-django
pillow
PyJWT==2.1.0